+++ lib/libxview/notify/ndet_g_sig.c
@@ -14,7 +14,9 @@ static char     sccsid[] = "@(#)ndet_g_sig.c 20.12 93/06/28 Copyr 1985 Sun Micro
  * Ndet_g_sig.c - Implement the notify_get_signal_func interface.
  */
 
-#include <xview_private/ntfy.h>
+#include <xview_private/ndet_g_sig_.h>
+#include <xview_private/ndetgetfnc_.h>
+#include <xview_private/ndet_sig_.h>
 #include <xview_private/ndet.h>
 
 extern          Notify_func
@@ -30,5 +32,5 @@ notify_get_signal_func(nclient, signal, mode)
 	return (NOTIFY_FUNC_NULL);
     if (ndet_check_sig(signal))
 	return (NOTIFY_FUNC_NULL);
-    return (ndet_get_func(nclient, type, (NTFY_DATA) signal, NTFY_USE_DATA));
+    return (ndet_get_func(nclient, type, (NTFY_DATA)(long)signal, NTFY_USE_DATA));
 }
