+++ lib/libxview/notify/ndet_s_fd.c
@@ -15,7 +15,14 @@ static char     sccsid[] = "@(#)ndet_s_fd.c 20.13 93/06/28 Copyr 1985 Sun Micro"
  * that are shared among NTFY_INPUT, NTFY_OUTPUT and NTFY_EXCEPTION.
  */
 
-#include <xview_private/ntfy.h>
+#include <xview_private/ndet_s_fd_.h>
+#include <xview_private/ndet_fd_.h>
+#include <xview_private/ndisdispch_.h>
+#include <xview_private/nint_set_.h>
+#include <xview_private/ntfyclient_.h>
+#include <xview_private/ntfy_cond_.h>
+#include <xview_private/ntfy_ctbl_.h>
+#include <xview_private/ntfyprotec_.h>
 #include <xview_private/ndet.h>
 #include <xview_private/ndis.h>
 #include <xview_private/nint.h>
@@ -41,7 +48,7 @@ ndet_set_fd_func(nclient, func, fd, type)
 	goto Done;
     /* Find/create condition */
     if ((condition = ntfy_new_condition(&(client->conditions), type,
-	     &(client->condition_latest), (NTFY_DATA) fd, NTFY_USE_DATA)) ==
+	     &(client->condition_latest), (NTFY_DATA)(long)fd, NTFY_USE_DATA)) ==
 	NTFY_CONDITION_NULL)
 	goto Done;
     ntfy_add_to_table(client, condition, type);
@@ -50,7 +57,7 @@ ndet_set_fd_func(nclient, func, fd, type)
     /* Remove condition if func is null */
     if (func == NOTIFY_FUNC_NULL) {
 	ndis_flush_condition(nclient, type,
-			     (NTFY_DATA) fd, NTFY_USE_DATA);
+			     (NTFY_DATA)(long)fd, NTFY_USE_DATA);
 	ntfy_unset_condition(&ndet_clients, client, condition,
 			     &ndet_client_latest, NTFY_NDET);
     }
