+++ lib/libxview/notify/ninti_wait.c
@@ -14,7 +14,9 @@ static char     sccsid[] = "@(#)ninti_wait.c 20.12 93/06/28 Copyr 1985 Sun Micro
  * Nint_i_wait.c - Implement the notify_interpose_wait3_func interface.
  */
 
-#include <xview_private/ntfy.h>
+#include <xview_private/ninti_wait_.h>
+#include <xview_private/ndet_wait_.h>
+#include <xview_private/nint_inter_.h>
 #include <xview_private/ndet.h>
 #include <xview_private/nint.h>
 
@@ -28,5 +30,5 @@ notify_interpose_wait3_func(nclient, func, pid)
     if (ndet_check_pid(pid))
 	return (notify_errno);
     return (nint_interpose_func(nclient, func, NTFY_WAIT3,
-				(NTFY_DATA) pid, NTFY_USE_DATA));
+				(NTFY_DATA)(long)pid, NTFY_USE_DATA));
 }
